Fix: Sub. IDs duplicated files
There are some duplicated files affecting
Subordinate IDs RPC and Utils. Those must
be added in a single place and keep the
existing functionality.
